+++ db/migrate/20190508200756_add_project_id_to_reports.rb
@@ -1,0 +1,6 @@
+class AddProjectIdToReports < ActiveRecord::Migration[6.0]
+  def change
+    add_column :reports, :project_id, :bigint
+    add_index :reports, :project_id
+  end
+end